requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
filelock==3.3.2
gensim==4.1.2
guppy3==3.1.2
huggingface-hub==0.0.19
idna==3.3
joblib==1.1.0
nltk==3.6.5
numpy==1.21.3
packaging==21.2
pandas==1.3.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
regex==2021.10.23
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
transformers==4.12.2
typing-extensions==3.10.0.2
urllib3==1.26.7
youtube-transcript-api==0.4.1
